Section 30 [Testing Node.js Applications]

+++ ex-3-rest-basics/server/controllers/feed/deletePost.js
@@ -1,3 +1,5 @@
+const { isValidObjectId } = require('mongoose');
+
 const Post = require('../../models/Post');
 const ErrorHandler = require('../../utils/ErrorHandler');
 const io = require('../../websocket');
@@ -6,6 +8,11 @@ const deletePost = async (req, res, next) => {
 	try {
 		const { user } = req;
 		const { id: postId } = req.params;
+
+		if (!isValidObjectId(postId)) {
+			throw new ErrorHandler(404, 'Post not found.');
+		}
+
 		await Post.findOneAndRemove({ _id: postId, creator: user._id }).orFail(
 			new ErrorHandler(403, 'Unauthorized user.')
 		);
